requirements.txt
certifi==2024.2.2
cffi==1.16.0
cryptography==42.0.5
gpsd-py3==0.3.0
jsonpickle==3.0.3
paho-mqtt==2.0.0
pycparser==2.21
pygnssutils==1.0.21
pynmea2==1.19.0
pynmeagps==1.0.33
pyrtcm==1.0.16
pyserial==3.5
pyspartn==0.2.0
pyubx2==1.2.39
